requirements.txt
selenium
webdriver_manager
requests
Pillow
